created a gradle project
