--- README.md
@@ -1,6 +1,0 @@
-# Ctrl-Alt-Elite, 420-2024fa
-Team Members: 
-- Sydney Norgaard
-- Jordan Rios
-- Sujan Gurung
-- Bishal Subedi
