+++ app/build.gradle
@@ -1,31 +1,21 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * This generated file contains a sample Java application project to get you started.
- * For more details on building Java & JVM projects, please refer to https://docs.gradle.org/8.8/userguide/building_java_projects.html in the Gradle documentation.
- */
-
 plugins {
-    // Apply the application plugin to add support for building a CLI application in Java.
     id 'application'
 }
 
 repositories {
-    // Use Maven Central for resolving dependencies.
     mavenCentral()
 }
 
 dependencies {
     // Use JUnit Jupiter for testing.
     testImplementation libs.junit.jupiter
-
     testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 
-    // This dependency is used by the application.
-    implementation libs.guava
+    // Dependencies used by the application.
+    implementation 'com.google.code.gson:gson:2.11.0'
+    implementation 'com.google.guava:guava:31.0.1-jre' // Specify the Guava version
 }
 
-// Apply a specific Java toolchain to ease working on different environments.
 java {
     toolchain {
         languageVersion = JavaLanguageVersion.of(21)
@@ -33,11 +23,22 @@ java {
 }
 
 application {
-    // Define the main class for the application.
-    mainClass = 'fa.ctrl.alt.elite.App'
+    mainClass = 'fa.ctrl.alt.elite.Main' // Ensure this points to your main class
 }
 
 tasks.named('test') {
-    // Use JUnit Platform for unit tests.
     useJUnitPlatform()
 }
+
+run {
+    standardInput = System.in
+}
+
+run {
+    doFirst {
+        println("Starting the application...")
+    }
+}
+
+
+
