+++ app/src/main/java/fa/ctrl/alt/elite/Main.java
@@ -1,3 +1,5 @@
+package fa.ctrl.alt.elite;
+
 import java.io.IOException;
 import java.util.Scanner;
 
@@ -141,7 +143,7 @@ public class Main {
                     umlEditor.listClass(listClassName);
                     break;
 
-                case "ListRelationships":
+                case "ListRlationships":
                     // Lists all relationships between classes
                     umlEditor.listRelationships();
                     break;
