+++ app/src/test/java/fa/ctrl/alt/elite/JsonUtilsTest.java
@@ -1,4 +1,4 @@
-package gradletest;
+package fa.ctrl.alt.elite;
 
 import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.AfterEach;
@@ -8,6 +8,10 @@ import org.junit.jupiter.api.Test;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Set;
 
 public class JsonUtilsTest {
